+++ src/mapping/pool.ts
@@ -1,4 +1,3 @@
-import { Address, BigDecimal, BigInt, ByteArray, Bytes } from "@graphprotocol/graph-ts"
 import {
   PoolAbi,
   NewAdmin,
@@ -7,22 +6,10 @@ import {
   RemoveLiquidity,
   AddMargin,
   RemoveMargin,
+  AddMarket,
 } from "../../generated/Pool/PoolAbi"
-import {
-  VaultAbi
-} from "../../generated/Pool/VaultAbi"
-import {
-  AavePoolAbi
-} from "../../generated/Pool/AavePoolAbi"
-import {
-  AaveOracleAbi
-} from "../../generated/Pool/AaveOracleAbi"
-import {
-  ERC20Abi
-} from "../../generated/Pool/ERC20Abi"
-import { getOrInitBToken, getOrInitDToken, getOrInitPool, getOrInitSymbolManager, getOrInitVault } from "../helpers/initializers"
-import { ZERO_ADDRESS } from "../utils/constants"
-import { handleLiquidityAction, handleMarginAction, initSymbols } from "./helper"
+import { getOrInitDToken, getOrInitPool, getOrInitSymbolManager } from "../helpers/initializers"
+import { handleLiquidityAction, handleMarginAction, initBTokens, initSymbols } from "./helper"
 import { formatDecimal } from "../utils/converters"
 
 export function handlePoolNewAdmin(event: NewAdmin): void {
@@ -58,41 +45,13 @@ export function handlePoolNewImplementation(event: NewImplementation): void {
   pool.save()
 
   // bToken init
-  const vaultContract = VaultAbi.bind(Address.fromBytes(pool.vaultImplementation))
-  const vault = getOrInitVault(pool.vaultImplementation)
-  vault.aavePool = vaultContract.aavePool()
-  vault.aaveOracle = vaultContract.aaveOracle()
-  vault.save()
-
-  const aavePoolContract = AavePoolAbi.bind(Address.fromBytes(vault.aavePool))
-  const aaveOracleContract = AaveOracleAbi.bind(Address.fromBytes(vault.aaveOracle))
-  const allAssets = aavePoolContract.getReservesList()
-  for (let i = 0; i < allAssets.length; i++) {
-    const asset = allAssets[i]
-    const configData = aavePoolContract.getConfiguration(asset).toHexString()
-    const market = (asset == pool.tokenB0) ? pool.marketB0 : (asset == pool.tokenWETH) ? pool.marketWETH : contract.markets(asset)
-
-    // ignore not support market
-    if (market == Bytes.fromHexString(ZERO_ADDRESS)) {
-      continue
-    } 
-    const bToken = getOrInitBToken(asset)
-    const bTokenContract = ERC20Abi.bind(Address.fromBytes(asset))
-    const marketContract = ERC20Abi.bind(Address.fromBytes(market))
-    bToken.bToken = asset
-    bToken.bTokenSymbol = bTokenContract.symbol()
-    bToken.bTokenDecimals = bTokenContract.decimals()
-    bToken.market = market
-    bToken.marketSymbol = marketContract.symbol()
-    bToken.marketDecimals = marketContract.decimals()
-    bToken.bTokenPrice = aaveOracleContract.getAssetPrice(asset).toBigDecimal()
-      .div(aaveOracleContract.BASE_CURRENCY_UNIT().toBigDecimal())
-    bToken.collateralFactor = BigInt.fromI32(I32.parseInt(configData.slice(configData.length - 4), 16)).toBigDecimal().div(BigDecimal.fromString("10000"))
-    bToken.pool = pool.id
-    bToken.save()
-  }
+  initBTokens(pool)
   // symbol init
   initSymbols(pool)
+}
+export function handleAddMarket(event: AddMarket): void {
+  const pool = getOrInitPool(event.address)
+  initBTokens(pool)
 }
 
 export function handlePoolAddLiquidity(event: AddLiquidity): void {