+++ changelog.md
@@ -5,3 +5,4 @@
 1. margin and liquidity need to get by contract call
 2. add account to position/liquidity/margin
 3. add txhash/timestamp is possible
+4. add addmarket/addSymbol/removeSymbol